terminat buton delete
+implementat masuri de securitate la stergerea unui user(sa nu poti sterge user-ul pe care esti conectat)

+++ includes/ajax/post_delete_users.php
@@ -6,14 +6,31 @@ chdir('../../');
 define("_ROOT", getcwd() . "/");
 
 include "includes/config.php";
-
-$id=$_POST["id"];
-
 global $db;
-try{
-    $stmt=$db->prepare("delete from user where id=:id");
-    $stmt->bindParam(':id',$id,PDO::PARAM_INT);
-    $stmt->execute();
-} catch (Exception $ex) {
-    die($ex->getMessage());
+$id_user=  get_logged_in_user()["id"];
+$multiple=$_POST["multiple"];
+if($multiple=="true"){
+    $id= $_POST["id"];
+    for($i=0;$i<count($id);$i++){
+        if($id[$i]!=$id_user){
+            try{
+                $stmt=$db->prepare("delete from user where id=:id");
+                $stmt->bindParam(':id',$id[$i],PDO::PARAM_INT);
+                $stmt->execute();
+            } catch (Exception $ex) {
+                die($ex->getMessage());
+            }
+        }
+    }
+}else{
+    $id= $_POST["id"];
+    if($id!=$id_user){
+        try{
+            $stmt=$db->prepare("delete from user where id=:id");
+            $stmt->bindParam(':id',$id,PDO::PARAM_INT);
+            $stmt->execute();
+        } catch (Exception $ex) {
+            die($ex->getMessage());
+        }
+    }
 }
